+++ cliboa/core/builder.py
@@ -16,10 +16,11 @@ import copy
 from cliboa.core.executor import _StepExecutor
 from cliboa.core.factory import _cliboa_factory, _get_scenario_loader_class
 from cliboa.core.interface import _IExecute
-from cliboa.core.listener import StepListener, StepStatusListener
 from cliboa.core.loader import _ScenarioLoader
 from cliboa.core.model import CommandArgument, ParallelStepModel, ScenarioModel, StepModel
 from cliboa.core.processor import _ParallelProcessor
+from cliboa.listener.base import BaseStepListener
+from cliboa.listener.step import StepStatusListener
 from cliboa.util.base import _BaseObject
 from cliboa.util.cache import StepArgument
 from cliboa.util.exception import CliboaException
@@ -114,7 +115,7 @@ class _ScenarioBuilder(_BaseObject):
         if isinstance(step, StepModel):
             instance = self._create_executor(step)
             # save arguments to refer by symbol as a global variables.
-            self._step_argument_cls._put(step.step, instance)
+            self._step_argument_cls._put(step.step, instance.step)
             return instance
         elif isinstance(step, ParallelStepModel):
             instances = []
@@ -122,7 +123,7 @@ class _ScenarioBuilder(_BaseObject):
                 instance = self._create_executor(p_step)
                 instances.append(instance)
                 # save arguments to refer by symbol as a global variables.
-                self._step_argument_cls._put(p_step.step, instance)
+                self._step_argument_cls._put(p_step.step, instance.step)
             return self._resolve(
                 "parallel_processor", _ParallelProcessor, instances, step.parallel_config
             )
@@ -142,7 +143,7 @@ class _ScenarioBuilder(_BaseObject):
             executor.register_listener(lis)
         return executor
 
-    def _create_listeners(self, step: StepModel) -> list[StepListener]:
+    def _create_listeners(self, step: StepModel) -> list[BaseStepListener]:
         listeners = [self._resolve("step_status_listener", StepStatusListener())]
         if step.listeners is not None:
             if step.arguments:
